Various bugfixes related to 1.38 changes. (#1109)
* Removed MDL uniform tags

* Removed unsafe conditional optimization.

* Allow thin-film layered over conductor (result is not correct, needs more work, but the setup is now valid).
