+++ libraries/pbrlib/genglsl/mx_conductor_bsdf.glsl
@@ -1,6 +1,6 @@
 #include "pbrlib/genglsl/lib/mx_microfacet_specular.glsl"
 
-void mx_conductor_bsdf_reflection(vec3 L, vec3 V, vec3 P, float occlusion, float weight, vec3 ior_n, vec3 ior_k, vec2 roughness, vec3 N, vec3 X, int distribution, out BSDF result)
+void mx_conductor_bsdf_reflection(vec3 L, vec3 V, vec3 P, float occlusion, float weight, vec3 ior_n, vec3 ior_k, vec2 roughness, vec3 N, vec3 X, int distribution, thinfilm tf, out BSDF result)
 {
     if (weight < M_FLOAT_EPS)
     {
@@ -18,7 +18,7 @@ void mx_conductor_bsdf_reflection(vec3 L, vec3 V, vec3 P, float occlusion, float
     float NdotH = clamp(dot(N, H), M_FLOAT_EPS, 1.0);
     float VdotH = clamp(dot(V, H), M_FLOAT_EPS, 1.0);
 
-    FresnelData fd = mx_init_fresnel_conductor(ior_n, ior_k);
+    FresnelData fd = tf.thickness > 0.0 ? mx_init_fresnel_conductor_airy(ior_n, ior_k, tf.thickness, tf.ior) : mx_init_fresnel_conductor(ior_n, ior_k);
     vec3 F = mx_compute_fresnel(VdotH, fd);
 
     float avgRoughness = mx_average_roughness(roughness);
@@ -31,7 +31,7 @@ void mx_conductor_bsdf_reflection(vec3 L, vec3 V, vec3 P, float occlusion, float
     result = D * F * G * comp * occlusion * weight / (4 * NdotV);
 }
 
-void mx_conductor_bsdf_indirect(vec3 V, float weight, vec3 ior_n, vec3 ior_k, vec2 roughness, vec3 N, vec3 X, int distribution, out BSDF result)
+void mx_conductor_bsdf_indirect(vec3 V, float weight, vec3 ior_n, vec3 ior_k, vec2 roughness, vec3 N, vec3 X, int distribution, thinfilm tf, out BSDF result)
 {
     if (weight < M_FLOAT_EPS)
     {
@@ -43,7 +43,7 @@ void mx_conductor_bsdf_indirect(vec3 V, float weight, vec3 ior_n, vec3 ior_k, ve
 
     float NdotV = clamp(dot(N, V), M_FLOAT_EPS, 1.0);
 
-    FresnelData fd = mx_init_fresnel_conductor(ior_n, ior_k);
+    FresnelData fd = tf.thickness > 0.0 ? mx_init_fresnel_conductor_airy(ior_n, ior_k, tf.thickness, tf.ior) : mx_init_fresnel_conductor(ior_n, ior_k);
     vec3 F = mx_compute_fresnel(NdotV, fd);
 
     vec3 Li = mx_environment_radiance(N, V, X, roughness, distribution, fd);
