+++ libraries/pbrlib/genosl/mx_conductor_bsdf.osl
@@ -1,6 +1,6 @@
 #include "pbrlib/genosl/lib/mx_microfacet_specular.osl"
 
-void mx_conductor_bsdf(float weight, color ior_n, color ior_k, vector2 roughness, normal N, vector U, string distribution, output BSDF result)
+void mx_conductor_bsdf(float weight, color ior_n, color ior_k, vector2 roughness, normal N, vector U, string distribution, thinfilm tf, output BSDF result)
 {
     if (weight < M_FLOAT_EPS)
     {
