--- source/MaterialXGenShader/Nodes/SheenBsdfNode.h
@@ -1,35 +1,0 @@
-//
-// TM & (c) 2017 Lucasfilm Entertainment Company Ltd. and Lucasfilm Ltd.
-// All rights reserved.  See LICENSE.txt for license.
-//
-
-#ifndef MATERIALX_SHEENBSDFNODE_H
-#define MATERIALX_SHEENBSDFNODE_H
-
-#include <MaterialXGenShader/Nodes/SourceCodeNode.h>
-#include <MaterialXGenShader/Nodes/HwSourceCodeNode.h>
-
-namespace MaterialX
-{
-
-/// Sheen BSDF node.
-class SheenBsdfNode : public SourceCodeNode
-{
-public:
-    static ShaderNodeImplPtr create();
-
-    void addInputs(ShaderNode& node, GenContext&) const override;
-};
-
-/// Sheen BSDF node specifically for HW.
-class HwSheenBsdfNode : public HwSourceCodeNode
-{
-  public:
-    static ShaderNodeImplPtr create();
-
-    void addInputs(ShaderNode& node, GenContext&) const override;
-};
-
-} // namespace MaterialX
-
-#endif
